# app/definition/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: OnJobs
  description: API de servicios de OnJobs
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /onjobs/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /users/:id/:
    get:
      tags:
        - Usuario
      description: Obtener usuario (ID de usuario)
      # used as the method name of the controller
      operationId: findById
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Usuario
      description: Actualizar usuario (ID de usuario)
      # used as the method name of the controller
      operationId: updateUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Usuario
      description: Eliminar usuario (ID de usuario)
      # used as the method name of the controller
      operationId: deleteUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /company/:
    get:
      tags:
        - Empresa
      description: Listar todas las empresas
      # used as the method name of the controller
      operationId: findAllCompanys
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Empresa
      description: Registrar empresa
      # used as the method name of the controller
      operationId: addCompany
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Company"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /company/:id/:
    get:
      tags:
        - Empresa
      description: Obtener empresa (ID de empresa)
      # used as the method name of the controller
      operationId: findCompanyById
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Empresa
      description: Actualizar empresa (ID de Empresa)
      # used as the method name of the controller
      operationId: updateCompany
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Company"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Empresa
      description: Eliminar empresa (ID de Empresa)
      # used as the method name of the controller
      operationId: deleteCompany
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /company/:id/login/:
    post:
      tags:
        - Empresa
      description: Login empresa
      # used as the method name of the controller
      operationId: companyLogin
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/CompanyLogin"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  User:
    type: object
    properties:
      email:
        description: Correo del usuario
        type: string
      name:
        description: Nombre
        type: string
      lastname:
        description: Apellido
        type: string
      rut:
        description: Rut
        type: number
      birthdate:
        description: Fecha de Nacimiento
        type: string
        format: date
      profession:
        description: Profesión
        type: string
      experience:
        description: Experiencia
        type: string
      region:
        description: Region
        type: number
      city:
        description: Ciudad
        type: number
      description:
        description: Descripción
        type: string
      flag:
        description: Recomienda
        type: boolean
      recomendation:
        description: Recomendación
        type: string
      score:
        description: Puntuación
        type: number
...
  Company:
    type: object
    properties:
      name:
        description: Nombre empresa
        type: string
      email:
        description: Email de empresa
        type: string
      password:
        description: Contraseña de empresa
        type: string
        format: password
      phone:
        description: Telefono de empresa
        type: number
      businessName:
        description: Razón Social
        type: string
      rut:
        description: Rut de empresa
        type: string
      entry:
        description: Rubro de la empresa
        type: string
      businessTurn:
        description: Giro de la empresa
        type: string
      region:
        description: Región
        type: string
      city:
        description: Ciudad
        type: string
      location:
        description: Dirección
        type: string
      flag:
        description: Recomienda
        type: boolean
      recomendation:
        description: Recomendación
        type: string
  CompanyLogin:
    type: object
    properties:
      email:
        description: Email de empresa
        type: string
      password:
        description: Contraseña
        type: string
        format: password
...
  OnJobsResponse:
    type: string
  ErrorResponse:
    type: string